feat: complete mystery box product update, image upload, and Supabase…

+++ BusinessObjects/ProductInMangaBox.cs
@@ -17,6 +17,6 @@ namespace BusinessObjects
         public string ProductId { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public float Chance { get; set; }
+        public double Chance { get; set; }
     }
 }
